movement is now handled with floats rather than 12.3 shorts to increase world size limit beyond -4096,4096

+++ engine/game/q_shared.h
@@ -481,8 +481,8 @@ typedef struct
 {
 	pmtype_t	pm_type;
 
-	short		origin[3];		// 12.3
-	short		velocity[3];	// 12.3
+	float		origin[3];		// float
+	float		velocity[3];	// float
 	byte		pm_flags;		// ducked, jump_held, etc
 	byte		pm_time;		// each unit = 8 ms
 	short		gravity;
